+++ app/static/lib/drag-selector.js
@@ -2,6 +2,13 @@ import {
   svgNameSpace
 } from './dom-utils.js'
 import * as att from './attribute-classes.js';
+import {
+  setCursorToId
+} from './utils.js';
+import {
+  cm,
+  platform
+} from './main.js';
 
 export function addDragSelector(v, vp) {
 
@@ -13,6 +20,7 @@ export function addDragSelector(v, vp) {
   var start = {};
   var end = {};
   var rect;
+  let firstElementHighlighted = false; // flag to highlight first selected element
 
   let noteSelector = '.note';
   let restSelector = '.rest,.mRest,.beatRpt,.halfmRpt,.mRpt';
@@ -23,7 +31,7 @@ export function addDragSelector(v, vp) {
   vp.addEventListener('mousedown', ev => {
     dragging = true;
     // clear selected elements, if no CMD/CTRL key is pressed
-    if (!((navigator.appVersion.indexOf("Mac") !== -1) && ev.metaKey) && !ev.ctrlKey) {
+    if (!(platform.startsWith('mac') && ev.metaKey) && !ev.ctrlKey) {
       v.selectedElements = [];
       v.updateHighlight();
     }
@@ -138,8 +146,13 @@ export function addDragSelector(v, vp) {
           if (yKeys) yKeys.forEach(yKey => {
             let els = obobj[xKey][yKey];
             if (els) els.forEach(e => {
-              if (!newEls.includes(e.id))
+              if (!newEls.includes(e.id)) {
                 newEls.push(e.id);
+                if (!firstElementHighlighted) {
+                  setCursorToId(cm, e.id);
+                  firstElementHighlighted = true;
+                }
+              }
             });
           });
         });
@@ -157,6 +170,7 @@ export function addDragSelector(v, vp) {
     if (svgPm && Array.from(svgPm.childNodes).includes(rect))
       svgPm.removeChild(rect);
     oldEls = [];
+    firstElementHighlighted = false;
   });
 
 }