+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Automatically load schema from provided `@meiversion` attribute, when no schema is given in <xml-model>
 * Filter settings through text input panel (incl. settings refactoring)
 * Annotation list layout fixes
+* Support autoValidate (true|false) as URL parameter
 ### 0.6.2 patch: small bug fixes
 * Bugfix: create valid <ptr/> objects for linking annotations
 * Bugfix: guard against edge-cases when switching vrvTk versions